refactor: remove prompt template file and enhance error handling in GptHaAssistant
